refactor(web): dedupe setState-in-render regression assertion into a test helper
Both CanvasThumb.test.tsx and BrowserLocalCanvasPage.test.tsx carried an
identical copy of the React setState-in-render warning regex and the
console.error-filter assertion (one inline, one as a local helper). Extract
them into src/test-utils/no-setstate-in-render.ts so the subtle regex has a
single source of truth and CanvasThumb uses the same named assertion instead
of an inline filter. Behavior-preserving; both suites still pass.
